requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==3.2.25
django-apscheduler==0.6.2
djangorestframework==3.15.1
idna==3.7
importlib-metadata==6.7.0
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzlocal==5.1
urllib3==2.0.7
xmltodict==0.13.0
zipp==3.15.0
